hide codeblocks for 500ms on initial load (#690)
* hide codeblocks for 500ms on initial load

code blocks render with dark theme initially and once javascript kicks
in, their theme is changed to the one set by user.

In case user chooses light theme, code blocks appear with dark theme for
just a few frames and only during initial load.

However, that is enough to create an odd "flashing" effect when
opening some page with code blocks.

An ideal fix would be to serve user a page with appropriate theme in the
first place, i.e. if user chooses light theme, then all the assets are
tailored for light theme (as opposed to current implementation, which is
that all assets are available always).

Unfortunately, this is not a trivial change, it would require some server
side logic, and currently there is server in play, it's all static files.

A non ideal workaround (this commit) is to hide code block elements for
500ms, give some time for their initial load, and then display.

Please note that this is only relevant for initial load. Once page is
loaded and user is navigating around by clicking buttons, none of this
is relevant, all themes work as expected.

* refactor delay script into a docusaurus plugin

* fix comment explaining the purpose of the plugin