docs: Update 'LATEST ON' page for better readability (#786)
* Update 2022-01-27-release-sql-jit-compiler.md

* Request for minor changes

When reading the 'LATEST ON keyword' page (https://questdb.io/docs/reference/sql/latest-on/), I was confused with one sentence. So, I updated it. If you feel it helps users, pls feel free to accept the change. 

Current: When LATEST ON is provided a single column is of type SYMBOL, the query will end as soon as all distinct symbol values have been found.
Updated: When `LATEST ON` is provided a single column, and it is of type SYMBOL, the query will end as soon as all distinct symbol values are found.

I made other minor sentence updates too. Pls feel free to accept/reject the changes, as per

* Update latest-on.md

Co-authored-by: Arijus Šukys <argshook@gmail.com>