cleanup of #790 (#885)
* add new tutorial tensorflow

* chore: fix cover image

* fix broken link

* chore: improvements

* fix: broken image

* fix: images

* fix: broken link

* rebuild

* fix: formatting and move to blog

* add new tutorial tensorflow

* chore: fix cover image

* fix broken link

* chore: improvements

* fix: broken image

* fix: images

* fix: broken link

* rebuild

* remove `.md` from tutorial, keep only blog post

* remove tutorial images

they exist under `static/img/blog/2022-06-07`

* move images from 2022-06-13 folder to 2022-06-07

to align with blog post publish date

* Fixing dependencies and indentation

For the tutorial to run I had to install pandas, numpy, scikit-learn, matplotlib, jupyter, and openpyxl. I also had to pip install tensorflow-macos instead of tensorflow.
I added the command to start the jupyter notebook
Fixed a misindented line that was causing an error

* Shortening a bit the intro and adding a filter to the query

I removed/reduced some paragraphs as the intro was a bit lengthy and readers probably want to get to the hands-on part.

To make questdb more relevant, I am selecting only a subset of the dataset. Otherwise we were just loading an xls file into questdb to then query the whole dataset, which fell a lot like reading directly from a CSV. With the filtering, we showcase that having a database is nice, and also that we have specific syntax to filter by year. I checked and we are still getting good enough predictions with 3 years worth of training data.

We would need to update also the referenced repository to reflect the query change and remove the data prep cell where we were removing zero values.

Since I am using a subset, we are not getting zero values anymore, so I removed directly that part of the data prep, making the article a bit shorter

* chore: small edits

* chore: minor fix

* fix: clean up images

Co-authored-by: Pei-Shan Wu <pswu@protonmail.com>
Co-authored-by: Maciej Bodek <maciej.bodek@gmail.com>
Co-authored-by: javier ramírez <javier@questdb.io>